Removing .toString on a byte array in AllowTrustOperation Builder Class
Executing a .toString() on a byte[] will result in a string representation of that byte[]'s memory address. This will cause the assetCode of an AllowTrustOperation that has an asset of asset type ASSET_TYPE_CREDIT_ALPHANUM12 to have an assetCode of some random address. This most certainly was not intended.