Remove Network singleton
Instead of using a static global variable to determine the network for a transaction, each
Server will be bound to a specific network. A transaction will also be bound to a network.
If you try to submit a transaction to a server which is bound to a different network an
exception will be raised.