Use bytes to represent memo text (#259)
The Java SDK assumes that all memo texts must be valid UTF 8 strings. However, that assumption is not valid. It turns out that any sequence of bytes is allowed to appear in the memo field (see stellar/go#2022 ).

Therefore, the correct representation for the memo field is a byte array.