Update XDR definitions for CAP 35 (#323)
As part of the work for #322 , I have updated the XDR definitions in the Java sdk to match those found in the go monorepo. I then ran xdrgen to generate the Java XDR classes. The only adjustment I had to do after generating the new XDR classes was to update src/main/java/org/stellar/sdk/AllowTrustOperation.java to handle the new representation of the asset field in the Allow Trust operation.